// src/types.ts
export type RnMutliSelectOption = {
    name: string,
    id: string,
    selected?: boolean,
    disabled?: boolean,
    isLabel?: boolean
    tagColor?: string
}

export interface RnMultiSelectProps {
    label: string;
    disable?: boolean,
    options: RnMutliSelectOption[];
    close?: Function;
    inputBoxStyles?: ViewStyle
    showSelectedTop?: boolean
    required?: boolean;
    triggerValidation?: boolean;
    selectedValues?: string,
    customOptionRender?: (item, index) => JSX.Element,
    showModal?: boolean,
    onChange?: Function,
    onClick?: Function,
    changeSearchText?: Function,
}